add specs and refactor

+++ app/controllers/questions_controller.rb
@@ -4,7 +4,13 @@ class QuestionsController < ApplicationController
   expose(:questions) { Question.all }
   expose :question
 
+  def show
+    @exposed_answer = Answer.new
+  end
+
   def create
+    @exposed_question = current_user.questions.new(question_params)
+
     if question.save
       redirect_to question, notice: 'Your question successfully created.'
     else
@@ -21,8 +27,12 @@ class QuestionsController < ApplicationController
   end
 
   def destroy
-    question.destroy
-    redirect_to questions_path
+    if current_user.author_of?(question)
+      question.destroy
+      redirect_to questions_path
+    else
+      redirect_to question_path(question)
+    end
   end
 
   private
@@ -30,5 +40,6 @@ class QuestionsController < ApplicationController
   def question_params
     params.require(:question).permit(:title, :body)
   end
-  
+
 end
+
